Implement caching private key in RSASignatureValidator
